+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # 2.61.8 (2023-06-05)
 
 ### Bug fixes
-- **date-picker**: Prevent `setDate` to be called infinitely when a date is set from the FormControl [mpellerin42]
+- **date-picker**: [mpellerin42]
+  - Prevent `setDate` to be called infinitely when a date is set from the FormControl
+  - Fix highlight of selected date
 
 # 2.61.7 (2023-06-02)